+++ test/test_aggregation.rb
@@ -30,7 +30,7 @@ class TestAggregation < TestHBaseJRubyBase
 
     # No projection
     assert_raise(ArgumentError) { @table.aggregate(:sum) }
-    assert_raise(ArgumentError) { @table.each.aggregate(:sum) }
+    assert_raise(ArgumentError) { @table.scoped.aggregate(:sum) }
 
     # Invalid type
     assert_raise(ArgumentError) { @table.project('cf1:a').aggregate(:sum, :float) }