+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+0.3.0
+-----
+- Easier data access with table schema
+  - `HBase::Table#schema=` method to provide the schema of the table
+  - You can omit column family names on predefined columns
+  - Automatic type conversion for known columns
+- *0.3.0 brings many backward-incompatible changes*
+  - *`Row#to_hash` and `Row#to_hash_with_versions` are now deprecated*. Use `to_h` and `to_H` instead without arguments.
+  - Default parameters for `HBase::Row#rowkey` and `HBase::Cell#rowkey` are now `:raw` instead of `:string`.
+  - `HBase::ColumnKey` is removed. Use plain 2-element Arrays instead.
+  - Enumerable classes (Table, Scoped, Row, ByteArray) now return Enumerator on each method when block not given
+  - `Cell#value` can now return the correct data type if defined in the schema. For Java byte array, use `Cell#raw`.
+  - `Row#[type|types]` methods no more take Array of columns as arguments
+- Added `HBase::Table#scoped` method to return `HBase::Scoped` object for the table
+- Added `HBase::Cell#{eql?,hash}` method for equaility check
+
 0.2.6
 -----
 - Fixed `HBase::Scoped#filter` method to take short integer (byte, short, int) values